refactor: extract useStats hook
